server/requirements.txt
Flask>=2.0.0 
flask-cors>=6.0.1
python-dotenv>=1.0.0
